Organize member chip NFT badge components.
