+++ packages/webapp/src/members/components/member-chip-components/index.ts
@@ -1,1 +1,2 @@
+export * from "./nft-info";
 export * from "./telegram-link";
