+++ packages/webapp/src/members/components/member-chip-components/telegram-link.tsx
@@ -22,7 +22,7 @@ export const MemberChipTelegramLink = ({
 
     return (
         <div onClick={goToTelegram} className={className}>
-            <p className="flex items-center text-xs text-gray-500 font-light hover:underline">
+            <p className="flex items-center text-xs text-gray-500 hover:text-gray-600 transition font-light hover:underline">
                 <FaTelegram className="mr-1" />
                 {parsedHandle}
                 <OpensInNewTabIcon size={8} className="mb-1.5" />
