+++ packages/webapp/src/members/components/member-chip.tsx
@@ -2,15 +2,12 @@ import React from "react";
 import { useRouter } from "next/router";
 import dayjs from "dayjs";
 
-import { atomicAssets, blockExplorerAccountBaseUrl } from "config";
+import { blockExplorerAccountBaseUrl } from "config";
 import { ROUTES } from "_app/routes";
-import { assetToLocaleString, openInNewTab, useCountdown } from "_app";
-import { GenericMemberChip } from "_app/ui";
-import { NFT } from "_app/ui/icons";
-import { MemberAuctionData } from "members";
+import { GenericMemberChip, openInNewTab } from "_app";
 
-import { MemberChipTelegramLink } from "./member-chip-components";
-import { AssetData, MemberData } from "../interfaces";
+import { MemberChipTelegramLink, NFTInfo } from "./member-chip-components";
+import { MemberData } from "../interfaces";
 
 interface MemberChipProps {
     member: MemberData;
@@ -69,83 +66,3 @@ const MemberDetails = ({ member, onClick }: MemberDetailsProps) => {
         </div>
     );
 };
-
-const NFTInfo = ({ member }: { member: MemberData }) => {
-    if (member.auctionData) {
-        return <AuctionBadge auctionData={member.auctionData} />;
-    }
-
-    if (!member.assetData) return null;
-
-    if (member.saleId) {
-        return (
-            <SaleBadge assetData={member.assetData} saleId={member.saleId} />
-        );
-    }
-
-    return <AssetBadge assetData={member.assetData} />;
-};
-
-const AuctionBadge = ({ auctionData }: { auctionData: MemberAuctionData }) => {
-    const countdown = useCountdown({
-        endTime: new Date(auctionData.bidEndTime as number),
-        interval: 30000,
-    });
-
-    const goToAuction = (e: React.MouseEvent) => {
-        e.stopPropagation();
-        const url = `${atomicAssets.hubUrl}/market/auction/${auctionData.auctionId}`;
-        openInNewTab(url);
-    };
-
-    return (
-        <div
-            className="flex items-center space-x-1 text-gray-500 hover:text-gray-600 transition"
-            onClick={goToAuction}
-        >
-            <NFT size={17} className="pb-px" />
-            {auctionData.price ? (
-                <>
-                    <p>{assetToLocaleString(auctionData.price, 4)}</p>
-                    <p>•</p>
-                </>
-            ) : null}
-            <p>{countdown["d-h-m"] || "auction ended"}</p>
-        </div>
-    );
-};
-
-const SaleBadge = ({
-    assetData,
-    saleId,
-}: {
-    assetData: AssetData;
-    saleId: string;
-}) => (
-    <div
-        className="flex items-center space-x-1 text-gray-500 hover:text-gray-600 transition"
-        onClick={(e) => {
-            e.stopPropagation();
-            const url = `${atomicAssets.hubUrl}/market/sale/${saleId}`;
-            openInNewTab(url);
-        }}
-    >
-        <NFT size={17} className="pb-px" />
-        <p>#{assetData.templateMint} ON SALE</p>
-    </div>
-);
-
-const AssetBadge = ({ assetData }: { assetData: AssetData }) => (
-    <div
-        onClick={(e) => {
-            e.stopPropagation();
-            const url = `${atomicAssets.hubUrl}/explorer/asset/${assetData?.assetId}`;
-            openInNewTab(url);
-        }}
-    >
-        <NFT
-            size={17}
-            className="pb-px text-gray-500 hover:text-gray-600 transition"
-        />
-    </div>
-);
